+++ cModule.py
@@ -12,6 +12,9 @@ class cModule(object):
     oModule.sUniqueName = sBinaryName.lower();
     oModule.sFileVersion = None;
     oModule.sTimestamp = None;
+    # The start address will have none of the upper 32 bits set on x86 for obvious reasons, but it will always have
+    # some bits set on x64 (citation needed).
+    oModule.sISA = uStartAddress < 0x100000000 and "x86" or "x64";
   
   def foGetOrCreateFunction(oModule, sSymbol):
     if sSymbol not in oModule._doFunction_by_sSymbol: